refactor(ai): extract remaining 6 confirm-handler dispatchers
Final Phase 0.5 batch. Extracts the six remaining case branches from
confirm/handler.ts into per-domain dispatcher modules, completing the
split begun in #125, #126, and #127. All nine cases now flow through
dispatchers/ with the `return await` pattern established in #127.

New dispatcher files:
- dispatchers/chat.ts            handles send_chat_message and
                                 send_group_chat_message (uses
                                 DirectChatSupabase / GroupChatSupabase
                                 type adaptation; chat-message case also
                                 persists a thread-recipient breadcrumb
                                 to ai_threads.metadata)
- dispatchers/discussions.ts     handles create_discussion_thread and
                                 create_discussion_reply
- dispatchers/enterprise-invites handles create_enterprise_invite and
                                 revoke_enterprise_invite (unique:
                                 create uses the auth-bound supabase
                                 client for the RPC, revoke uses
                                 serviceSupabase for the UPDATE; both
                                 are threaded through
                                 EnterpriseInviteDispatcherContext)

No behaviour change. All nine case branches are now 10-15 line
dispatcher invocations wrapping an identical ctx + deps shape.

handler.ts: 677 -> 356 LOC (-321 lines). Total across full Phase 0.5:
920 -> 356 (-564 lines, 61% reduction).

Quality gates:
- 21/21 characterization tests in pending-actions-handler.test.ts pass
- 272/272 AI route tests pass (all of tests/routes/ai/*.ts)
- tsc + eslint clean

Stacks on top of #127 (feat/ai-confirm-handler-split-jobs). Phase 0.5
is now complete; subsequent phases can parallelize safely because
handler.ts is a thin dispatch shell.
